frontend/requirements.txt
streamlit
openai
python-dotenv
azure-cosmos==4.6.0  
audio_recorder_streamlit
papermill